Port Plurr to Rust 🦀
The implementation tries to stay as close to the original code as
possible, although due to the nature of the language, some constructs
are not possible, and different idioms/expressions must be used, e.g.
pattern matching via `match`.

This also affects the API, because in Rust there's no way (yet) to
provide optional arguments, and string-formatting parameters are
therefore fed via `.param("key", "value")` pairs. To keep things simple,
this only accepts strings for the time being.

Likewise, a macro could be written to support a `format` call with
variadic arguments, but that's left as a potential future improvement.

The tests include all syntax features, and also potential formatting
errors.